Znakomstvo s Python/lesson_4-Nariezaiem_avatarki/requirements.txt
pillow == 7.0.0
